requirements.txt
python-docx==1.2.0
lxml==6.0.2
boto3